[update] ManipulateListSegmentクラスに定義してあったメソッドを全てShortFormatクラス内のprivateに定義。 LongFormatはShortFormatを継承するように変更
